resolve comments 2
